Modernize rewrite API and Rails integration
