--- lib/generators/templates/uploadcare_config_template.erb
@@ -1,165 +1,0 @@
-# frozen_string_literal: true
-
-Uploadcare::Rails.configure do |config|
-  # =============================================================================
-  # API Keys
-  # =============================================================================
-
-  # Sets your Uploadcare public key (required for File Uploader)
-  config.public_key = ENV.fetch('UPLOADCARE_PUBLIC_KEY', 'demopublickey')
-
-  # Sets your Uploadcare secret key (required for REST API operations)
-  # config.secret_key = ENV.fetch('UPLOADCARE_SECRET_KEY', 'demosecretkey')
-
-  # =============================================================================
-  # File Storage Options
-  # =============================================================================
-
-  # Stores files to Uploadcare servers after object save
-  config.store_files_after_save = true
-
-  # Deletes files from Uploadcare servers after object destroy
-  config.delete_files_after_destroy = true
-
-  # These options enable ActiveJob background jobs to perform file operations asynchronously.
-  # IMPORTANT: Enabling these options requires a configured ActiveJob adapter (e.g., Sidekiq,
-  # Resque, or Delayed Job) and a working queueing system. Without a proper ActiveJob backend,
-  # enabling these options will cause runtime failures.
-  # Note: store_files_async only takes effect when store_files_after_save is true.
-  # Note: delete_files_async only takes effect when delete_files_after_destroy is true.
-  config.store_files_async = false
-  config.delete_files_async = false
-
-  # =============================================================================
-  # Caching Options
-  # =============================================================================
-
-  # Sets caching for Uploadcare files
-  config.cache_namespace = 'uploadcare'
-  config.cache_files = true
-  config.cache_expires_in = 1.day
-
-  # =============================================================================
-  # File Uploader Configuration
-  # See https://uploadcare.com/docs/file-uploader/configuration/ for details
-  # =============================================================================
-
-  # If true, only image files are allowed to be uploaded.
-  # config.img_only = false
-
-  # Allow multiple file uploads
-  # config.multiple = false
-
-  # Sets the list of file types that can be uploaded.
-  # Example: 'image/*,video/*,.pdf'
-  # config.accept = nil
-
-  # Sets the list of preferred MIME types for external sources.
-  # Example: 'image/*,video/*'
-  # config.external_sources_preferred_types = nil
-
-  # Defines if files should be stored permanently after upload.
-  # 'auto' - respect project settings
-  # 'true' - always store
-  # 'false' - never store
-  # config.store = 'auto'
-
-  # Mirrors the camera preview horizontally
-  # config.camera_mirror = false
-
-  # Sets upload sources available in the File Uploader.
-  # Example: 'local,url,camera,dropbox,gdrive'
-  # config.source_list = 'local,url,camera,dropbox,gdrive'
-
-  # Maximum file size in bytes for local uploads
-  # config.max_local_file_size_bytes = nil
-
-  # Sets your custom CDN CNAME
-  # config.cdn_cname = nil
-
-  # Sets custom Upload API base URL
-  # config.base_url = 'https://upload.uploadcare.com'
-
-  # Sets custom social sources base URL
-  # config.social_base_url = 'https://social.uploadcare.com'
-
-  # =============================================================================
-  # Security Options
-  # =============================================================================
-
-  # Secure signature for signed uploads (requires secret key on server)
-  # config.secure_signature = nil
-
-  # Signature expiration time
-  # config.secure_expire = nil
-
-  # Proxy URL for secure delivery
-  # config.secure_delivery_proxy = nil
-
-  # =============================================================================
-  # Upload Performance Options
-  # =============================================================================
-
-  # Maximum number of retries for throttled requests
-  # config.retry_throttled_request_max_count = 1
-
-  # Minimum file size for multipart uploads (in bytes)
-  # config.multipart_min_file_size = 26_214_400
-
-  # Chunk size for multipart uploads (in bytes)
-  # config.multipart_chunk_size = 5_242_880
-
-  # Maximum concurrent upload requests
-  # config.max_concurrent_requests = 10
-
-  # Maximum attempts for multipart upload
-  # config.multipart_max_attempts = 3
-
-  # =============================================================================
-  # URL Upload Options
-  # =============================================================================
-
-  # Check for URL duplicates before uploading
-  # config.check_for_url_duplicates = false
-
-  # Save URLs for recurrent uploads
-  # config.save_url_for_recurrent_uploads = false
-
-  # =============================================================================
-  # Output Options
-  # =============================================================================
-
-  # Output file group instead of individual files when multiple is enabled
-  # config.group_output = true
-
-  # =============================================================================
-  # Localization
-  # =============================================================================
-
-  # Available locales: ar, ca, cs, da, de, el, en, es, et, fr, he, hr, hu, id,
-  # it, ja, ko, nb, nl, pl, pt, ro, ru, sk, sr, sv, tr, uk, vi, zh-CN, zh-TW
-  config.locale = 'en'
-
-  # Custom metadata to attach to uploaded files
-  # config.metadata = {}
-
-  # =============================================================================
-  # Image Options
-  # =============================================================================
-
-  # Remove Uploadcare copyright from the image editor
-  # config.remove_copyright = false
-
-  # Preset for image cropping. Example: '4:3', '1:1', 'free'
-  # config.crop_preset = nil
-
-  # Shrink images on client before uploading. Example: '800x600'
-  # config.image_shrink = nil
-
-  # =============================================================================
-  # CDN Configuration
-  # =============================================================================
-
-  # Defines your CDN hostname for serving files
-  config.cdn_hostname = 'ucarecdn.com'
-end
