+++ lib/generators/uploadcare_config_generator.rb
@@ -1,14 +1,11 @@
 # frozen_string_literal: true
 
-# A generator that copies the config template to config directory
 class UploadcareConfigGenerator < Rails::Generators::Base
-  desc "This generator creates an uploadcare config file your_app/config/initializers/uploadcare.rb"
+  desc "This generator creates a config/uploadcare.yml file for uploadcare-rails"
 
   source_root File.expand_path("templates", __dir__)
 
-  # Copies uploadcare initializer template into host app.
-  # @return [void]
   def create_config_file
-    copy_file "uploadcare_config_template.erb", "config/initializers/uploadcare.rb"
+    copy_file "uploadcare_config.yml.erb", "config/uploadcare.yml"
   end
 end
