+++ lib/uploadcare/rails.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 require "uploadcare/rails/version"
+require "uploadcare-rails"
 
 module Uploadcare
   module Rails
-    # Base error class for uploadcare-rails runtime failures.
     class Error < StandardError; end
   end
 end
