+++ lib/uploadcare/rails/action_view/uploadcare_include_tags.rb
@@ -81,4 +81,3 @@ module Uploadcare
   end
 end
 
-ActionView::Base.include Uploadcare::Rails::ActionView::UploadcareIncludeTags
