+++ lib/uploadcare/rails/internal/active_storage_uuid_mapping.rb
@@ -1,14 +1,13 @@
 # frozen_string_literal: true
 
-module ActiveStorage
-  class Service
-    class UploadcareService < Service
-      # UUID lookup and persistence helpers for key-to-file mapping.
-      module UuidMapping
+module Uploadcare
+  module Rails
+    module Internal
+      module ActiveStorageUuidMapping
         private
 
         def uuid_for!(key)
-          uuid_for(key) || raise(ActiveStorage::FileNotFoundError)
+          uuid_for(key) || raise(::ActiveStorage::FileNotFoundError)
         end
 
         def uuid_for(key)
@@ -21,16 +20,16 @@ module ActiveStorage
         end
 
         def uuid_from_blob(key)
-          return unless defined?(ActiveStorage::Blob)
+          return unless defined?(::ActiveStorage::Blob)
 
-          blob = ActiveStorage::Blob.find_by(key: key)
+          blob = ::ActiveStorage::Blob.find_by(key: key)
           blob&.metadata&.[]("uploadcare_uuid")
         end
 
         def persist_uuid_to_blob(key, uuid)
-          return unless defined?(ActiveStorage::Blob)
+          return unless defined?(::ActiveStorage::Blob)
 
-          blob = ActiveStorage::Blob.find_by(key: key)
+          blob = ::ActiveStorage::Blob.find_by(key: key)
           return unless blob
 
           metadata = (blob.metadata || {}).dup
@@ -41,17 +40,17 @@ module ActiveStorage
         end
 
         def keys_for_prefix(prefix)
-          if defined?(ActiveStorage::Blob)
+          if defined?(::ActiveStorage::Blob)
             sanitized_prefix = sanitize_sql_like_prefix(prefix)
-            return ActiveStorage::Blob.where("key LIKE ?", "#{sanitized_prefix}%").pluck(:key)
+            return ::ActiveStorage::Blob.where("key LIKE ?", "#{sanitized_prefix}%").pluck(:key)
           end
 
           @key_uuid_map.keys.select { |key| key.start_with?(prefix) }
         end
 
         def sanitize_sql_like_prefix(prefix)
-          if defined?(ActiveRecord::Base) && ActiveRecord::Base.respond_to?(:sanitize_sql_like)
-            ActiveRecord::Base.sanitize_sql_like(prefix.to_s)
+          if defined?(::ActiveRecord::Base) && ::ActiveRecord::Base.respond_to?(:sanitize_sql_like)
+            ::ActiveRecord::Base.sanitize_sql_like(prefix.to_s)
           else
             prefix.to_s.gsub(/[\\%_]/) { |char| "\\#{char}" }
           end
