--- lib/uploadcare/rails/mongoid/mount_uploadcare_file.rb
@@ -1,85 +1,0 @@
-# frozen_string_literal: true
-
-require "mongoid"
-require "active_support/concern"
-require "uploadcare/rails/services/id_extractor"
-require "uploadcare/rails/jobs/delete_file_job"
-require "uploadcare/rails/jobs/store_file_job"
-require "uploadcare/rails/objects/file"
-
-module Uploadcare
-  module Rails
-    module Mongoid
-      module MountUploadcareFile
-        extend ActiveSupport::Concern
-
-        def build_uploadcare_file(attribute, client: nil)
-          cdn_url = read_attribute(attribute).to_s
-          return if cdn_url.empty?
-
-          uuid = IdExtractor.call(cdn_url)
-          cache_key = File.build_cache_key(cdn_url)
-          default_attributes = { cdn_url: cdn_url, uuid: uuid.presence }
-          file_attributes = ::Rails.cache.read(cache_key).presence || default_attributes
-          Uploadcare::Rails::File.new(file_attributes, client: client)
-        end
-
-        module ClassMethods
-          def has_uploadcare_file(attribute, uploadcare_client: nil)
-            define_method attribute do
-              client = resolve_uploadcare_client(uploadcare_client)
-              build_uploadcare_file(attribute, client: client)
-            end
-
-            define_method "uploadcare_store_#{attribute}!" do |store_job = StoreFileJob|
-              file = public_send(attribute)
-              return unless file&.uuid
-
-              client = resolve_uploadcare_client(uploadcare_client)
-              if Uploadcare::Rails.configuration.store_files_async
-                client_options = client ? Uploadcare::Rails.serialize_client_options(client) : {}
-                return store_job.perform_later(file.uuid, client_options)
-              end
-
-              (client || Uploadcare::Rails.client).files.batch_store(uuids: [file.uuid])
-            end
-
-            define_method "uploadcare_delete_#{attribute}!" do |delete_job = DeleteFileJob|
-              file = public_send(attribute)
-              return unless file&.uuid
-
-              client = resolve_uploadcare_client(uploadcare_client)
-              if Uploadcare::Rails.configuration.delete_files_async
-                client_options = client ? Uploadcare::Rails.serialize_client_options(client) : {}
-                return delete_job.perform_later(file.uuid, client_options)
-              end
-
-              (client || Uploadcare::Rails.client).files.batch_delete(uuids: [file.uuid])
-            end
-
-            if Uploadcare::Rails.configuration.store_files_after_save
-              set_callback(:save, :after, :"uploadcare_store_#{attribute}!", if: :"#{attribute}_changed?")
-            end
-
-            if Uploadcare::Rails.configuration.delete_files_after_destroy
-              set_callback(:destroy, :after, :"uploadcare_delete_#{attribute}!")
-            end
-          end
-
-          alias_method :mount_uploadcare_file, :has_uploadcare_file
-        end
-
-        private
-
-        def resolve_uploadcare_client(client_source)
-          return nil if client_source.nil?
-          return instance_exec(&client_source) if client_source.respond_to?(:call)
-
-          client_source
-        end
-      end
-    end
-  end
-end
-
-Mongoid::Document.include Uploadcare::Rails::Mongoid::MountUploadcareFile
