+++ spec/uploadcare/rails/action_view/uploadcare_form_builder_spec.rb
@@ -1,14 +1,13 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-require 'uploadcare/rails/action_view/uploadcare_uploader_tags'
-require 'uploadcare/rails/action_view/uploadcare_form_builder'
+require 'uploadcare/rails/internal/uploader_field_helpers'
+require 'uploadcare/rails/action_view/form_builder'
 
-RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :helper do
+RSpec.describe 'ActionView::Helpers::FormBuilder uploadcare helpers', type: :helper do
   let(:public_key) { 'demopublickey' }
   let(:config_attributes) { { pubkey: public_key } }
 
-  # Mock errors class
   let(:mock_errors_class) do
     Class.new do
       def initialize(errors = {})
@@ -21,7 +20,6 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
     end
   end
 
-  # Mock model class
   let(:model_class) do
     errors_class = mock_errors_class
     Struct.new(:avatar, :errors, keyword_init: true) do
@@ -34,22 +32,20 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
   before do
     allow(Uploadcare::Rails).to receive_message_chain(:configuration, :uploader_config_attributes)
       .and_return(config_attributes)
-    # Use a fixed UUID for predictable test output
     allow(SecureRandom).to receive(:uuid).and_return('test-uuid-1234')
   end
 
-  # Helper to create a FormBuilder instance
   def form_builder_for(object, object_name: nil)
     object_name ||= object.class.name.underscore rescue 'object'
     ActionView::Helpers::FormBuilder.new(object_name, object, self, {})
   end
 
-  describe '#uploadcare_file' do
+  describe '#uploadcare_file_field' do
     it 'generates all uploader components' do
       user = model_class.new
       builder = form_builder_for(user, object_name: 'user')
 
-      tag = builder.uploadcare_file(:avatar)
+      tag = builder.uploadcare_file_field(:avatar)
 
       expect(tag).to include('<uc-form-input')
       expect(tag).to include('<uc-config')
@@ -61,7 +57,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
       user = model_class.new
       builder = form_builder_for(user, object_name: 'user')
 
-      tag = builder.uploadcare_file(:avatar)
+      tag = builder.uploadcare_file_field(:avatar)
 
       expect(tag).to include('name="user[avatar]"')
     end
@@ -70,7 +66,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
       user = model_class.new
       builder = form_builder_for(user, object_name: 'user')
 
-      tag = builder.uploadcare_file(:avatar)
+      tag = builder.uploadcare_file_field(:avatar)
 
       expect(tag).to include('ctx-name="test-uuid-1234"')
     end
@@ -79,7 +75,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
       user = model_class.new
       builder = form_builder_for(user, object_name: 'user')
 
-      tag = builder.uploadcare_file(:avatar, ctx_name: 'custom-uploader')
+      tag = builder.uploadcare_file_field(:avatar, ctx_name: 'custom-uploader')
 
       expect(tag).to include('ctx-name="custom-uploader"')
     end
@@ -90,7 +86,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
           user = model_class.new
           builder = form_builder_for(user, object_name: 'user')
 
-          tag = builder.uploadcare_file(:avatar, solution: solution)
+          tag = builder.uploadcare_file_field(:avatar, solution: solution)
 
           expect(tag).to include("<uc-file-uploader-#{solution}")
         end
@@ -101,7 +97,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
       user = model_class.new
       builder = form_builder_for(user, object_name: 'user')
 
-      tag = builder.uploadcare_file(:avatar, multiple: true, img_only: true)
+      tag = builder.uploadcare_file_field(:avatar, multiple: true, img_only: true)
 
       expect(tag).to include('multiple="true"')
       expect(tag).to include('img-only="true"')
@@ -112,7 +108,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
         user = model_class.new(errors: mock_errors_class.new({}))
         builder = form_builder_for(user, object_name: 'user')
 
-        tag = builder.uploadcare_file(:avatar)
+        tag = builder.uploadcare_file_field(:avatar)
 
         expect(tag).not_to include('field_with_errors')
       end
@@ -123,7 +119,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
         user = model_class.new(errors: mock_errors_class.new(avatar: [ "can't be blank" ]))
         builder = form_builder_for(user, object_name: 'user')
 
-        tag = builder.uploadcare_file(:avatar)
+        tag = builder.uploadcare_file_field(:avatar)
 
         expect(tag).to include('field_with_errors')
         expect(tag).to include('<uc-form-input')
@@ -133,9 +129,8 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
         user = model_class.new(errors: mock_errors_class.new(avatar: [ "can't be blank" ]))
         builder = form_builder_for(user, object_name: 'user')
 
-        tag = builder.uploadcare_file(:avatar)
+        tag = builder.uploadcare_file_field(:avatar)
 
-        # The entire output should be wrapped
         expect(tag).to start_with('<div class="field_with_errors">')
         expect(tag).to include('<uc-config')
         expect(tag).to include('<uc-file-uploader-regular')
@@ -147,7 +142,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
         user = model_class.new(errors: mock_errors_class.new(email: [ 'is invalid' ]))
         builder = form_builder_for(user, object_name: 'user')
 
-        tag = builder.uploadcare_file(:avatar)
+        tag = builder.uploadcare_file_field(:avatar)
 
         expect(tag).not_to include('field_with_errors')
       end
@@ -157,7 +152,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
       it 'does not wrap the field with error markup' do
         builder = ActionView::Helpers::FormBuilder.new('user', nil, self, {})
 
-        tag = builder.uploadcare_file(:avatar)
+        tag = builder.uploadcare_file_field(:avatar)
 
         expect(tag).not_to include('field_with_errors')
       end
@@ -165,7 +160,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
       it 'still generates uploader components' do
         builder = ActionView::Helpers::FormBuilder.new('user', nil, self, {})
 
-        tag = builder.uploadcare_file(:avatar)
+        tag = builder.uploadcare_file_field(:avatar)
 
         expect(tag).to include('<uc-form-input')
         expect(tag).to include('<uc-file-uploader-regular')
@@ -177,10 +172,23 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
         plain_object = Object.new
         builder = form_builder_for(plain_object, object_name: 'item')
 
-        tag = builder.uploadcare_file(:avatar)
+        tag = builder.uploadcare_file_field(:avatar)
 
         expect(tag).not_to include('field_with_errors')
       end
+    end
+  end
+
+  describe '#uploadcare_files_field' do
+    it 'passes multiple and group_output automatically' do
+      user = model_class.new
+      builder = form_builder_for(user, object_name: 'user')
+
+      tag = builder.uploadcare_files_field(:avatar)
+
+      expect(tag).to include('multiple="true"')
+      expect(tag).to include('group-output="true"')
+      expect(tag).to include('name="user[avatar]"')
     end
   end
 
@@ -199,7 +207,7 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
       user = model_class.new(errors: mock_errors_class.new(avatar: [ "can't be blank" ]))
       builder = form_builder_for(user, object_name: 'user')
 
-      tag = builder.uploadcare_file(:avatar)
+      tag = builder.uploadcare_file_field(:avatar)
 
       expect(tag).to include('<span class="custom-error">')
       expect(tag).not_to include('field_with_errors')
@@ -208,6 +216,6 @@ RSpec.describe 'ActionView::Helpers::FormBuilder#uploadcare_file_field', type: :
 end
 
 RSpec.configure do |c|
-  c.include Uploadcare::Rails::ActionView::UploadcareUploaderTags, type: :helper
+  c.include Uploadcare::Rails::Internal::UploaderFieldHelpers, type: :helper
   c.include ActionView::Helpers::FormHelper, type: :helper
 end
